# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DokMart - Digital document marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDocumentsResponse"
...
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a single document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/admin/create:
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /levels:
    get:
      operationId: listLevels
      tags: [levels]
      summary: List all levels
      responses:
        "200":
          description: List of levels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListLevelsResponse"
    post:
      operationId: createLevel
      tags: [levels]
      summary: Create a new level (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLevelBody"
      responses:
        "201":
          description: Created level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"
...
  /levels/{id}:
    delete:
      operationId: deleteLevel
      tags: [levels]
      summary: Delete a level (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        level:
          type: string
        semester:
          type: ["string", "null"]
        docType:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: integer
        previewUrl:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        pageCount:
          type: ["integer", "null"]
        downloadCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - subject
        - level
        - price
        - isFeatured
        - downloadCount
        - createdAt
    ListDocumentsResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - documents
        - total
        - page
        - totalPages
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        level:
          type: string
        semester:
          type: ["string", "null"]
        docType:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: integer
        previewUrl:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
          default: false
        pageCount:
          type: ["integer", "null"]
      required:
        - title
        - description
        - subject
        - level
        - price
    UpdateDocumentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        level:
          type: string
        semester:
          type: ["string", "null"]
        docType:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: integer
        previewUrl:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        pageCount:
          type: ["integer", "null"]
...
    Level:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        group:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        documentCount:
          type: integer
      required:
        - id
        - name
        - slug
        - sortOrder
        - documentCount
    ListLevelsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Level"
    CreateLevelBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        group:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
          default: 0
      required:
        - name
        - slug
...